Refactering for models
