+++ app/controllers/application_controller.rb
@@ -1,9 +1,10 @@
 class ApplicationController < ActionController::Base
-  # Authentication
-  before_action :authenticate_user!
-
+  # Include
   include CommonError
   include DeviseManager
   include AdminUserAuthentication
   include UserRolePolicy
+
+  # Authentication
+  before_action :authenticate_user!
 end
