+++ app/models/advertisement.rb
@@ -6,8 +6,8 @@ class Advertisement < ApplicationRecord
   validates :start_date, presence: true
   validates :end_date, presence: true
   validates :placed_shop, presence: true
-  validate :start_date_check
-  validate :end_date_check
+  validates_with StartDateValidator
+  validates_with EndDateValidator
 
   # AdContract -> Shop
   has_many :ad_contract, dependent: :delete_all
@@ -15,22 +15,4 @@ class Advertisement < ApplicationRecord
 
   # Nest Create
   accepts_nested_attributes_for :ad_contract, allow_destroy: true
-
-  private
-
-    def start_date_check
-      if self.start_date.present?
-        if self.start_date < Date.today
-          errors.add(:start_date, "に今日以降の日付を入力して下さい")
-        end
-      end
-    end
-
-    def end_date_check
-      if self.start_date.present? and self.end_date.present?
-        if self.start_date > self.end_date
-          errors.add(:end_date, "に開始日以降の日付を入力して下さい")
-        end
-      end
-    end
 end
