+++ app/models/application_record.rb
@@ -1,3 +1,5 @@
 class ApplicationRecord < ActiveRecord::Base
+  include ActiveModel::Validations
+
   self.abstract_class = true
 end
