+++ app/models/category.rb
@@ -1,3 +1,4 @@
 class Category < ApplicationRecord
+  # Tag
   has_many :tag
 end
