+++ app/models/color.rb
@@ -1,3 +1,4 @@
 class Color < ApplicationRecord
+  # User
   has_many :user
 end
