+++ app/models/concerns/profile_update_module.rb
@@ -1,0 +1,17 @@
+module ProfileUpdateModule
+  extend ActiveSupport::Concern
+
+  # Config for Updating Profile
+  def update_without_current_password(params, *options)
+    params.delete(:current_password)
+
+    if params[:password].blank? && params[:password_confirmation].blank?
+      params.delete(:password)
+      params.delete(:password_confirmation)
+    end
+
+    result = update_attributes(params, *options)
+    clean_up_passwords
+    result
+  end
+end
