+++ app/models/interest.rb
@@ -1,4 +1,5 @@
 class Interest < ApplicationRecord
+  # User <-> Tag
   belongs_to :user
   belongs_to :tag
 end
