+++ app/models/role.rb
@@ -1,14 +1,10 @@
 class Role < ApplicationRecord
+  # Validation
+  validates :resource_type, inclusion: { in: Rolify.resource_types }, allow_nil: true
+
+  # User
   has_and_belongs_to_many :users, :join_table => :users_roles
-
-  belongs_to :resource,
-            :polymorphic => true,
-            :optional => true
-
-
-  validates :resource_type,
-            :inclusion => { :in => Rolify.resource_types },
-            :allow_nil => true
+  belongs_to :resource, polymorphic: true, optional: true
 
   scopify
 end
