+++ app/models/sns_credential.rb
@@ -1,3 +1,4 @@
 class SnsCredential < ApplicationRecord
+  # User
   belongs_to :user
 end
