+++ app/models/tag.rb
@@ -1,6 +1,8 @@
 class Tag < ApplicationRecord
+  # Interest -> User
   has_many :interest, dependent: :delete_all
   has_many :user, through: :interest
 
+  # Category
   belongs_to :category
 end
